// src/app/modules/security/authorization/enums/security-authorization-enums.model.ts
/**********************************************************************************************
Code generated with MKL Plug-in version: 47.7.13
Code generated at time stamp: 2020-01-07T20:34:08.364
Copyright: Kerubin - logokoch@gmail.com

WARNING: DO NOT CHANGE THIS CODE BECAUSE THE CHANGES WILL BE LOST IN THE NEXT CODE GENERATION.
***********************************************************************************************/

import { SelectItem } from 'primeng/api';

export interface OrderStatus extends SelectItem {
	
  label?: string;
  value: any;
  
}

export interface PaymentMethod extends SelectItem {
	
  label?: string;
  value: any;
  
}

export interface AccountType extends SelectItem {
	
  label?: string;
  value: any;
  
}